refactor: Song, VotingSong 분리
