--- backend/src/main/java/shook/shook/part/application/dto/PartCommentRegisterRequest.java
@@ -1,14 +1,0 @@
-package shook.shook.part.application.dto;
-
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor(access = AccessLevel.PRIVATE)
-@AllArgsConstructor
-@Getter
-public class PartCommentRegisterRequest {
-
-    private String content;
-}
