--- backend/src/main/java/shook/shook/part/application/dto/PartRegisterRequest.java
@@ -1,23 +1,0 @@
-package shook.shook.part.application.dto;
-
-import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Positive;
-import jakarta.validation.constraints.PositiveOrZero;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor(access = AccessLevel.PRIVATE)
-@AllArgsConstructor
-@Getter
-public class PartRegisterRequest {
-
-    @NotNull
-    @PositiveOrZero
-    private Integer startSecond;
-
-    @NotNull
-    @Positive
-    private Integer length;
-}
