--- backend/src/main/java/shook/shook/part/exception/PartCommentException.java
@@ -1,32 +1,0 @@
-package shook.shook.part.exception;
-
-public class PartCommentException extends RuntimeException {
-
-    public static class NullOrEmptyPartCommentException extends PartCommentException {
-
-        public NullOrEmptyPartCommentException() {
-            super();
-        }
-    }
-
-    public static class TooLongPartCommentException extends PartCommentException {
-
-        public TooLongPartCommentException() {
-            super();
-        }
-    }
-
-    public static class CommentForOtherPartException extends PartCommentException {
-
-        public CommentForOtherPartException() {
-            super();
-        }
-    }
-
-    public static class DuplicateCommentExistException extends PartCommentException {
-
-        public DuplicateCommentExistException() {
-            super();
-        }
-    }
-}
