--- backend/src/main/java/shook/shook/song/domain/repository/dto/SongTotalVoteCountDto.java
@@ -1,10 +1,0 @@
-package shook.shook.song.domain.repository.dto;
-
-import shook.shook.song.domain.Song;
-
-public interface SongTotalVoteCountDto {
-
-    Song getSong();
-
-    Long getTotalVoteCount();
-}
