+++ backend/src/main/java/shook/shook/voting_song/domain/repository/VoteRepository.java
@@ -1,8 +1,8 @@
-package shook.shook.part.domain.repository;
+package shook.shook.voting_song.domain.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-import shook.shook.part.domain.Vote;
+import shook.shook.voting_song.domain.Vote;
 
 @Repository
 public interface VoteRepository extends JpaRepository<Vote, Long> {
