+++ backend/src/main/java/shook/shook/voting_song/exception/VoteException.java
@@ -1,4 +1,4 @@
-package shook.shook.part.exception;
+package shook.shook.voting_song.exception;
 
 public class VoteException extends RuntimeException {
