+++ backend/src/test/resources/test.sql
@@ -1,20 +1,43 @@
-drop table part;
 drop table song;
-drop table vote;
+drop table killing_part;
+drop table killing_part_comment;
 drop table voting_song;
 drop table voting_song_part;
-drop table register;
+drop table vote;
+drop table member;
 
-create table part
+create table if not exists song
+(
+    id              bigint auto_increment,
+    title           varchar(100) not null,
+    singer          varchar(50)  not null,
+    length          integer      not null,
+    video_url       text         not null,
+    album_cover_url text         not null,
+    created_at      timestamp(6) not null,
+    primary key (id)
+    );
+
+create table if not exists killing_part
 (
     id           bigint auto_increment,
     start_second integer      not null,
     length       varchar(255) not null check (length in ('SHORT', 'STANDARD', 'LONG')),
-    song_id      bigint,
+    song_id      bigint       not null,
     created_at   timestamp(6) not null,
     primary key (id)
-);
-create table song
+    );
+
+create table if not exists killing_part_comment
+(
+    id              bigint auto_increment,
+    killing_part_id bigint       not null,
+    content         varchar(200) not null,
+    created_at      timestamp(6) not null,
+    primary key (id)
+    );
+
+create table if not exists voting_song
 (
     id              bigint auto_increment,
     title           varchar(100) not null,
@@ -24,41 +47,31 @@ create table song
     album_cover_url text         not null,
     created_at      timestamp(6) not null,
     primary key (id)
-);
-create table vote
-(
-    id         bigint auto_increment,
-    part_id    bigint,
-    created_at timestamp(6) not null,
-    primary key (id)
-);
-create table voting_song_part
+    );
+create table if not exists voting_song_part
 (
     id             bigint auto_increment,
     start_second   integer      not null,
     length         varchar(255) not null check (length in ('SHORT', 'STANDARD', 'LONG')),
-    voting_song_id bigint,
+    voting_song_id bigint       not null,
     created_at     timestamp(6) not null,
     primary key (id)
-);
-create table voting_song
-(
-    id              bigint auto_increment,
-    title           varchar(100) not null,
-    singer          varchar(50)  not null,
-    length          integer      not null,
-    video_url       text         not null,
-    album_cover_url text         not null,
-    created_at      timestamp(6) not null,
-    primary key (id)
-);
-create table register
+    );
+create table if not exists vote
 (
     id                  bigint auto_increment,
-    voting_song_part_id bigint,
+    voting_song_part_id bigint       not null,
     created_at          timestamp(6) not null,
     primary key (id)
-);
+    );
+
+create table if not exists member
+(
+    id       bigint auto_increment,
+    email    varchar(100) not null,
+    nickname varchar(100) not null,
+    primary key (id)
+    );
 
 TRUNCATE TABLE song;
